--- src/main/java/com/github/can019/BaseApplication.java
@@ -1,17 +1,0 @@
-package com.github.can019;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.boot.web.servlet.ServletComponentScan;
-
-@SpringBootApplication
-@ConfigurationPropertiesScan("com.example.base.config")
-@ServletComponentScan
-public class BaseApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BaseApplication.class, args);
-	}
-
-}
