+++ src/main/java/com/github/can019/base/BaseApplication.java
@@ -1,0 +1,17 @@
+package com.github.can019.base;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.boot.web.servlet.ServletComponentScan;
+
+@SpringBootApplication
+@ConfigurationPropertiesScan("com.example.base.config")
+@ServletComponentScan
+public class BaseApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BaseApplication.class, args);
+	}
+
+}
